requirements.txt
accelerate>=0.16.0
torchvision
transformers>=4.25.1
ftfy
tensorboard
datasets
numpy
scipy
diffusers
rembg
trimesh
opencv-contrib-python
tqdm
wandb
simple_parsing
einops
imageio[ffmpeg]
ray["default"]
pandas
easydict
git+https://github.com/naver/croco/#subdirectory=models/curope
git+https://github.com/iamNCJ/dust3r.git
